infrastructure/cloud_functions/requirements.txt
# Cloud Functions Requirements
# These dependencies are needed for GCP Cloud Functions deployment

# Core
python-dotenv==1.0.0
pydantic==2.5.0

# AI/ML
openai==1.6.0
anthropic==0.7.0
tiktoken==0.5.2
pinecone-client==3.0.0

# Logging
structlog==23.2.0

# Google Cloud
google-cloud-pubsub==2.18.0
google-cloud-storage==2.14.0
functions-framework==3.5.0
